Fix bug order
