+++ server/src/routes/me.js
@@ -12,7 +12,7 @@ import {
 } from '../app/controllers/me.js';
 import {} from '../app/controllers/user.js';
 
-import { getUserOrders, getUserOrderById } from '../app/controllers/me.js';
+import { getUserOrders, getUserOrderById, cancelOrder } from '../app/controllers/me.js';
 
 const router = Router();
 
@@ -25,6 +25,8 @@ router.patch('/change-password', changePassword);
 router.get('/wishlist', getWishlist);
 router.post('/wishlist/:productId', addToWishlist);
 router.patch('/wishlist/:productId', removeFromWishlist);
-router.get('orders', getUserOrders);
-router.get('orders/:orderId', getUserOrderById);
+router.get('/orders', getUserOrders);
+router.get('/orders/:orderId', getUserOrderById);
+router.patch('/orders/:orderId/cancel', cancelOrder);
+
 export default router;
