+++ server/src/routes/order.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { deleteOrder, getOrder, getOrders, updateOrder ,checkOutOrder, updateOrderStatus,cancelOrder,deleteOrderByAdmin,getAllOrder,getOrderById,getUserOrderById } from '../app/controllers/order.js';
+import { deleteOrder, getOrder, getOrders, updateOrder ,checkOutOrder, updateOrderStatus,getAllOrder,getOrderById } from '../app/controllers/order.js';
 import { verifyAdmin } from '../app/middlewares/auth.js';
 
 const router = Router();
@@ -8,10 +8,9 @@ router.put('/:id', verifyAdmin, updateOrder);
 router.delete('/:id', verifyAdmin, deleteOrder);
 router.get('/:id', verifyAdmin, getOrder);
 router.get('/', verifyAdmin, getOrders);
-router.post('/orders', checkOutOrder);
-router.patch('/orders/:orderId/cancel', updateOrderStatus);
-router.delete('/orders/:orderId', verifyAdmin, cancelOrder)
-router.get('/orders',verifyAdmin, getAllOrder)
-router.get('/orders/:orderId', verifyAdmin,getOrderById)
+router.post('/', checkOutOrder);
+router.patch('/:orderId',verifyAdmin, updateOrderStatus);
+router.get('/',verifyAdmin, getAllOrder)
+router.get('/:orderId', verifyAdmin,getOrderById)
 
 export default router;
